+++ app/models/metric.rb
@@ -95,7 +95,7 @@ class Metric < ActiveRecord::Base
     evmethods = getEvmethods
     itemWeights = []
     unless evmethods.blank?
-      itemsLength = evmethods.map{|ev| ev.module.constantize.getItemsWithType("integer").length}.sum
+      itemsLength = evmethods.map{|ev| ev.module.constantize.getItemsWithType("numeric").length}.sum
       itemWeight = 1/itemsLength.to_f
       itemsLength.times do |i|
         itemWeights.push(BigDecimal(itemWeight,6))
